Server/PoPsyServer/requirements.txt
dj-database-url==0.5.0
Django==3.2.1
gunicorn==20.1.0
psycopg2==2.8.6
whitenoise==5.2.0
djangorestframework==3.12.4
mysqlclient==2.0.3
numpy~=1.20.3
scikit-learn~=0.24.2
jsonpickle~=2.0.0
dj-config-url~=0.1.1
tensorflow~=2.5.0rc3
googletrans~=3.0.0